Change operator type matrices
In this scheme, `-` is supposed to be the default sequencing operation. It
should be able to connect

+ source - pipe   ->  source
+ pipe   - pipe   ->  pipe
+ pipe   - sink   ->  sink
+ source - sink   ->  ready

`-` should convert callables into pipes that map the callable over the stream:

+ source - func   ->  source
+ pipe   - func   ->  pipe
+ func   - pipe   ->  pipe
+ func   - sink   ->  sink

To use callables in filters rather than maps, use `+` instead of `-`. This
should work in two situations:

+ source + func   ->  source
+ pipe   + func   ->  pipe

To use callables in sinks rather than pipes, use `>>` instead of `-`. This
should also work in two situations:

+ source >> func  ->  ready
+ pipe   >> func  ->  sink

Additionally `>>` is also being expected to work in two more situations:

+ source >> sink  ->  ready
+ func   >> sink  ->  sink

but these could be suppressed with no loss of generality.