Add xfail test and implementation for arg[a,b]
I don't see how this can be made to work, because the item getting syntax seems
to have no means of distinguishing between a tuple multiple values!

Consider

  class A:
      def __call__   (self, *args): return args
      def __getitem__(self, *args): return args

  a = A()

With this definition, we get

  a( 1,2 ) ->  (1,2)
  a((1,2)) -> ((1,2),)  # different from previous

  a[ 1,2 ] -> ((1,2),)
  a[(1,2)] -> ((1,2),)  # same as previous

In other words, multiple values inside [] are automatically wrapped in a tuple:
__getitem__ will never get more than one argument.

Consequently __getitem__ cannot distinguish between multiple values and a
single-value-that-is-a-tuple.

You can check see this in action in dictionares:

   d = {(1,2): "BINGO"}
   d[ 1,2 ] # -> 'BINGO'
   d[(1,2)] # -> 'BINGO'

It looks fundamentally broken.