Stop using futures for collecting results
In the original scheme, `coroutine_and_outputs` returned the coroutine along
with a sequence of outputs, consisting of name-future pairs.

In this scheme, rather than returning outputs, `coroutine_and_outputs` accepts
an additional parameter: a mutable container of outputs. When the coroutine
closes, rather than setting a future, it now adds a name-value pair to the
outputs container.

In the old scheme, the order of outputs was determined by the traversal order at
coroutine construction time. In this scheme it is determined by the order in
which the coroutines close. The former is more stable.

Additionally, in the old scheme, mutations are confined inside the futures; in
the new scheme, the outputs namespace is mutated all over the place.