Revert .travis.yml for master branch
The book needs Rust (and cargo cache) to build, while the package itself needs
python. I don't see how to have a sensible multi-language environment on Travis,
so, for now at least, there will be a version of .travis.yml on master, for
normal builds (essentially testing), and a different one on gh-pages for
publishing of the documentation.

As this work is about to be merged to master, the gh-pages version of
.travis.yml has to be reverted to the master version.